+++ src/transform/src/types.ts
@@ -1,7 +1,6 @@
 export class ProgramInfo {
   functions: FunctionSignature[];
   types: TypeDefinition[];
-  embedders: FunctionSignature[];
 }
 
 export class FunctionSignature {